service_recognize/requirements.txt
aioamqp==0.14.0
boto3==1.13.15
botocore==1.16.17
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.15.2
future==0.18.2
gevent==20.5.2
google-api-core==1.17.0
google-auth==1.15.0
google-cloud-speech==1.3.2
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.29.0
idna==2.9
jmespath==0.10.0
pamqp==2.3.0
pika==1.1.0
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==7.0.2
python-dateutil==2.8.1
python-ffmpeg==1.0.11
loguru==0.5.3
pytz==2020.1
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
sentry-sdk==0.14.4
six==1.15.0
urllib3==1.25.9
websockets==8.1
websocket-client==0.57.0
zope.event==4.4
zope.interface==5.1.0
